+++ src/main/java/com/learning/mfscreener/models/portfolio/UserTransactionDTO.java
@@ -3,15 +3,16 @@ package com.learning.mfscreener.models.portfolio;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record UserTransactionDTO(
         LocalDate date,
         String description,
-        Double amount,
-        Double units,
-        Double nav,
-        Double balance,
+        BigDecimal amount,
+        BigDecimal units,
+        BigDecimal nav,
+        BigDecimal balance,
         TransactionType type,
         @JsonProperty("dividend_rate") String dividendRate)
         implements Serializable {}
